feat: merge accs on auth
