+++ src/lib/utilities/api.js
@@ -1,5 +1,5 @@
 import { get } from "svelte/store";
-import { appStateStore } from "../../stores/state.js";
+import { appStateStore, userStateStore } from "../../stores/state.js";
 import { API_URL } from "../../configs/env.js";
 import { convertMetersToRadian, getCookie } from "./helpers.js";
 
@@ -51,11 +51,13 @@ const fetchFunction = async ({ url, method, credentials, headers, body }) => {
 
 const saveToDB = async (coords, rating, averageRating, comment, isPersonalExperience, timeline) => {
 	const url = `${API_URL}/geo/add`;
+	const { lang } = get(userStateStore);
 
 	return await fetchFunction({
 		url,
 		method: "POST",
 		body: JSON.stringify({
+			lang,
 			properties: {
 				rating,
 				averageRating,
@@ -116,8 +118,9 @@ const reactOnRating = async (ratingID, shouldReport) => {
 
 // POIs
 
-const savePOIToDB = async (coords, props, lang) => {
+const savePOIToDB = async (coords, props) => {
 	const url = `${API_URL}/poi/add`;
+	const { lang } = get(userStateStore);
 
 	return await fetchFunction({
 		url,
